пересобран компонент VacancyStatus
