+++ src/pages/vacancy/VacancyEdit/components/VacancyStatus/VacancyStatus.vue
@@ -41,90 +41,33 @@
 
   <Teleport to="body">
     <!-- Вывод модалки для добавления и изменения статуса -->
-    <Modal
-      class="vacancy-edit__modal"
-      :show="indicators.isAdd || indicators.isEdit"
-      @click.self="
+    <VacancyStatusModal
+        v-if="indicators.isAdd || indicators.isEdit"
+        :statusMod="statusMod"
+        :indicators="indicators"
+        :colors="colors"
+        :handleModification="handleModification"
+        @closeModal="
         indicators.isAdd = false;
         indicators.isEdit = false;
-      "
-    >
-      <template v-slot:header>
-        <div class="vacancy-edit__modal__title">
-          {{ indicators.isAdd ? 'Добавление статуса' : 'Изменение статуса' }}
-        </div>
-      </template>
-
-      <template v-slot:body>
-        <InputSimple
-          v-if="indicators.isAdd"
-          v-model="statusMod.name"
-          id="statusEdit"
-          inputType="input"
-          placeholder="Название статуса"
-        />
-        <InputSimple
-          v-model="statusMod.comment"
-          id="statusEdit"
-          inputType="textarea"
-          placeholder="Комментарий"
-        />
-
-        <div
-          v-if="statusMod.action !== 'delete'"
-          class="vacancy-edit__modal-select"
-        >
-          <div>Цвет:</div>
-          <select
-            v-model="statusMod.color"
-            class="vacancy-edit__modal-select-color"
-            :style="{ backgroundColor: statusMod.color }"
-          >
-            <option
-              v-for="color in colors"
-              :key="color.value"
-              :value="color.value"
-              :style="{ backgroundColor: color.value }"
-            >
-              {{ color.value }}
-            </option>
-          </select>
-        </div>
-
-        <ButtonMain
-          class="vacancy-edit__modal-btn-add"
-          @click="
-            indicators.isAdd
-              ? handleModification(statusMod.name, 'create')
-              : handleModification(statusMod.name, 'update')
-          "
-          :isActive="request"
-          :message="errorMessage"
-        >
-          <template v-slot:text>{{
-            indicators.isAdd ? 'Добавить' : 'Изменить'
-          }}</template>
-        </ButtonMain>
-      </template>
-    </Modal>
+         "
+    />
   </Teleport>
 </template>
 
 <script setup>
-import ButtonMain from '@/components/ButtonMain.vue';
 import {
   VacanciesGetVacancyStatuses,
   VacanciesModifyVacancyStatus,
   VacanciesSetVacancyStatusTransfer,
-} from '../js/ApiClassesVacancyEdit';
+} from '../../js/ApiClassesVacancyEdit.js';
 import { onMounted, ref, watch } from 'vue';
-import InputSimple from '@/components/InputSimple.vue';
 import ButtonIcon from '@/components/ButtonIcon.vue';
 import IconAdd from '@/assets/icons/add.svg?component';
-import Modal from '@/components/Modal.vue';
-import { colors } from '../js/statusColors.js';
+import { colors } from '../../js/statusColors.js';
 import VacancyStatusList from './VacancyStatusList.vue';
 import SpinnerMain from '@/components/SpinnerMain.vue';
+import VacancyStatusModal from './VacancyStatusModal.vue';
 const props = defineProps({
   // Id вакансии
   vacancyId: {
@@ -331,34 +274,10 @@ watch(
   transition: all 0.3s ease;
   width: 20px;
   height: 20px;
+
   &:hover {
     transform: scale(1.3);
   }
-}
-
-.vacancy-edit__modal-btn-add {
-  display: flex;
-  align-items: center;
-  align-self: center;
-  width: fit-content;
-  gap: 5px;
-}
-
-.vacancy-edit__modal-select {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 5px;
-}
-
-.vacancy-edit__modal-select-color {
-  -webkit-appearance: none;
-  appearance: none;
-  width: 90px;
-  height: 20px;
-  cursor: pointer;
-  border-radius: 10px;
-  text-align: center;
 }
 
 .vacancy-edit__status-spinner {
